fix: resolve all clippy warnings
- Remove unused variables (signing_key, my_onion)
- Remove unused SETTINGS_KEYS constant
- Rename MessageType::from_str to from_string (clippy confusing_with)
- Add Default impl for VchatState and WebRTCState
- Allow too_many_arguments clippy lint in store.rs
- Remove unnecessary usize cast in protocol.rs
- Move .await calls before DB lock in store.rs (Send fix)

+++ src-tauri/src/messaging/protocol.rs
@@ -296,7 +296,7 @@ pub fn deserialize_wire_message(data: &[u8]) -> Result<WireMessage> {
         );
     }
 
-    if frame_len > MAX_FRAME_SIZE as usize {
+    if frame_len > MAX_FRAME_SIZE {
         bail!(
             "Frame length {} exceeds MAX_FRAME_SIZE {}",
             frame_len,
